// src/app/models/contractModels/BillingModels/ServiceSearchModel.ts
export class ServiceSearchModel {

    serviceId: number;
    isActiveServiceList: boolean;
    cashAmount: number;
    departmentId: number;
    serviceName: string;
    providerId: number;
    serviceCode: string;
    insServiceCode: string;
    insServiceName: string;
    insDiscountAmount: number;
    insDiscountType: string;
    deptDepartmentName: string;
    deptDiscountType: string;
    deptDiscountAmount: number;
    patientShare: number;
    shareType: string;
    serviceDiscountAmount: number;
    patientShareAmount: number;
    insShareAmount: number;
    netShareAmount: number;
}